% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{adjust}
\alias{adjust}
\title{Geom adjustments}
\usage{
adjust(geom, field, value)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{field}{A character string specifying a data field.}

\item{value}{One or more of "stack", "dodge", "jitter", or "symmetric"
specifying a position adjustment for a field or entire geom.}
}
\description{
Describe an adjust for a geom.
}
\seealso{
Other geom methods: \code{\link{color}},
  \code{\link{opacity}}, \code{\link{position}},
  \code{\link{shape}}, \code{\link{size}}
}
