% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Create a G2 chart}
\usage{
chart(data, width = 600, height = 300)
}
\arguments{
\item{data}{A data frame.}

\item{width}{The width of the chart in pixels.}

\item{height}{The height of the chart in pixels}
}
\description{
A chart.
}
\examples{
games \%>\%
  chart() \%>\%
  interval(
    . \%>\%
      position("genre*sold") \%>\%
      color("genre")
  )

}
