% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-facet.R
\name{facet}
\alias{facet}
\title{Facet a chart}
\usage{
facet()
}
\arguments{
\item{type}{One of "rect", "list", "tree", "mirror", or "matrix" specifying
the type or style of faceting.

分面类型，现支持的类型为：rect, list, tree, mirror, matrix。}

\item{fields}{A character vector specifying the fields to facet on.

设置分面的字段，用于划分数据集。}

\item{showTitle}{TRUE or FALSE, defaults to TRUE.

是否显示分面的标题，默认为 true，即展示。}

\item{autoSetAxis}{TRUE or FALSE, defaults to TRUE.

是否自动设置坐标轴的文本，避免重复和遮挡，默认为 true，即自动设置。}

\item{padding}{A number or numeric vector.

设置每个 view 之间的间距。padding 是view 的内部边距，所以不会影响布局。}

\item{eachView}{A function of two arguments \code{view} and \code{facet}, returns
\code{NULL}.

回调函数，用于绘制每一个分面对应 view。该回调函数的内容如下：}

\item{colTitle}{A list specifying styles for column titles,\preformatted{colTitle: {
  offsetY: -15, // 列标题垂直方向的偏移
  style: {
    fontSize: 14,
    textAlign: 'center',
    fill: '#444'
  } // 标题文本样式
}
}}

\item{rowTitle}{A list specifying styles for row titles,\preformatted{rowTitle: {
  offsetX: -15, // 列标题水平方向的偏移
  style: {
    fontSize: 14,
    textAlign: 'center',
    fill: '#444'
  } // 标题文本样式
}
}}
}
\description{
Chart facet.
}
