% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golly.R
\docType{package}
\name{golly}
\alias{golly}
\alias{golly-package}
\alias{golly-package}
\title{golly, compose G2 visualizations}
\description{
Much like ggplot2, G2 is an implementation of the grammar of graphics. This
package provides a set of tools to build G2 visualizations.
}
\seealso{
\link{https://github.com/antvis/g2}
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nate@zevross.com}

}
