% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{interval}
\alias{interval}
\title{Interval geom}
\usage{
interval(chart, f)
}
\arguments{
\item{chart}{A chart object.}

\item{f}{A functional sequence of geom methods to apply to the interval
geom.}
}
\description{
Interval geoms are bar charts.
}
