% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.chart}
\alias{knit_print.chart}
\title{Knit a chart}
\usage{
knit_print.chart(x, ...)
}
\arguments{
\item{x}{A chart object.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Put a chart into a R markdown document.
}
