% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{opacity}
\alias{opacity}
\title{Geom opacity}
\usage{
opacity(geom, field, value = NULL)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{field}{A character string specifying a data field.}

\item{value}{A number between 0 and 1 specifying the opacity to apply to
the entire geom or field.}
}
\description{
Describe an opacity for a geom.
}
\seealso{
Other geom methods: \code{\link{adjust}},
  \code{\link{color}}, \code{\link{position}},
  \code{\link{shape}}, \code{\link{size}}
}
