% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{position}
\alias{position}
\title{Geom position}
\usage{
position(geom, value)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{value}{A character string specifying an interaction of fields or
a character vector specifying an interaction of fields.}
}
\description{
Describe how fields interact.
}
\details{
The "fieldA*fieldB" notation seems equivalent to "\link{'fieldA', 'fieldB'}"
notation.
}
\seealso{
\link{https://antv.alipay.com/zh-cn/g2/3.x/api/geom.html#_position}

Other geom methods: \code{\link{adjust}},
  \code{\link{color}}, \code{\link{opacity}},
  \code{\link{shape}}, \code{\link{size}}
}
