% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.chart}
\alias{print.chart}
\alias{as.character.chart}
\alias{as.tags.chart}
\title{Print a chart object}
\usage{
\method{print}{chart}(x, ...)

\method{as.character}{chart}(x, ...)

\method{as.tags}{chart}(x, ...)
}
\arguments{
\item{x}{A chart object.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Print a chart object or coerce a chart to a set of tags or HTML string.
}
