% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{shape}
\alias{shape}
\title{Geom shape}
\usage{
shape(geom, field, value = NULL)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{field}{A character string specifying a data field.}

\item{value}{A character string, unnamed character vector, or named character
vector or list. (as with \link{color}, a function can be supplied in native G2,
however as of now there is no good way to translate R functions to
JavaScript functions).

The following shapes are available for each geom,

\strong{point}: "hollowCircle", "hollowSquare", "hollowDiamond", "hollowBowtie",
"hollowTriangle", "hollowHexagon", "cross", "tick", "plus", "hyphen", and
"line"

\strong{line}: "line", "smooth", "dash", "dot", and "spline"

\strong{area}: "area", "line", and "smoothLine"

\strong{interval}: "rect", "hollowRect", "line", and "tick"

\strong{polygon}: "polygon" and "hollow"

\strong{schema}: "box" and "candle"}
}
\description{
Describe a shape for an entire geom or for certain fields.
}
\seealso{
Other geom methods: \code{\link{adjust}},
  \code{\link{color}}, \code{\link{opacity}},
  \code{\link{position}}, \code{\link{size}}
}
