% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{size}
\alias{size}
\title{Geom size}
\usage{
size(geom, field, value = NULL)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{field}{A character string specifying a data field.}

\item{value}{A number specifying a size or a vector of two numbers specifying
a min and max value to map field values between.}
}
\description{
Describe a size for a geom.
}
\seealso{
Other geom methods: \code{\link{adjust}},
  \code{\link{color}}, \code{\link{opacity}},
  \code{\link{position}}, \code{\link{shape}}
}
