% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{color}
\alias{color}
\title{Geom color}
\usage{
color(geom, value)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{value}{A character string (color name or hexcode), a gradiant specified
as \code{"color1-color2"}, an unnamed vector of colors to use, or a named vector
or list mapping values to colors.
(G2 also supports functions, if an R function could be converted to a JS
one).}
}
\description{
Describe a colors for fields.
}
