% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-methods.R
\name{position}
\alias{position}
\title{Geom position}
\usage{
position(geom, x, y)
}
\arguments{
\item{geom}{A geom object, e.g. a call to \link{line}, \link{point}, etc.}

\item{x}{The field on the x-axis.}

\item{y}{The field on the y-axis.}
}
\description{
Describe how fields interact.
}
\details{
The "fieldA*fieldB" notation seems equivalent to "\link{'fieldA', 'fieldB'}"
notation.
}
\seealso{
\link{https://antv.alipay.com/zh-cn/g2/3.x/api/geom.html#_position}
}
